Add generation.
Allows character by character generation. This includes the actual
generation code as well as changes to the model and data loading to
allow it. Generation code is still untested, but runs. Also, this could
cause a regression in the training code (specifically if it fails to
zero the hiddens, something which I believe is accounted for but hasn't
been tested).